requirements.txt
yt-dlp
requests
streamlit
faiss-cpu
sentence-transformers
pandas
groq
crawl4ai